moved all session related api calls into one file
